feat(ampd): add event filtering to event handlers (#1049)
