+++ ampd/src/grpc/reqs.rs
@@ -1,5 +1,3 @@
-use std::str::FromStr;
-
 use ampd_proto::{
     Algorithm, BroadcastRequest, ContractStateRequest, ContractsRequest, KeyId, KeyRequest,
     SignRequest, SubscribeRequest,
@@ -7,13 +5,13 @@ use ampd_proto::{
 use axelar_wasm_std::chain::ChainName;
 use axelar_wasm_std::nonempty;
 use cosmrs::Any;
-use error_stack::{bail, ensure, report, Report, Result, ResultExt};
-use report::ResultCompatExt;
+use error_stack::{bail, ensure, report, Result, ResultExt};
 use thiserror::Error;
 use tonic::Request;
 
-use crate::types::TMAddress;
-use crate::{tofnd, PREFIX};
+use crate::event_sub::event_filter::EventFilters;
+use crate::tofnd;
+use crate::types::AxelarAddress;
 
 type ContractQuery = Vec<u8>;
 type KeyIdString = nonempty::String;
@@ -34,7 +32,8 @@ impl Validate for Request<SubscribeRequest> {
             include_block_begin_end,
         } = self.into_inner();
 
-        (filters, include_block_begin_end).try_into()
+        EventFilters::try_from((filters, include_block_begin_end))
+            .change_context(Error::InvalidFilter)
     }
 }
 
@@ -49,7 +48,7 @@ impl Validate for Request<BroadcastRequest> {
 }
 
 impl Validate for Request<ContractStateRequest> {
-    type Output = (TMAddress, ContractQuery);
+    type Output = (AxelarAddress, ContractQuery);
 
     fn validate(self) -> Result<Self::Output, Error> {
         let ContractStateRequest { contract, query } = self.into_inner();
@@ -58,19 +57,12 @@ impl Validate for Request<ContractStateRequest> {
         let _: serde_json::Value =
             serde_json::from_slice(&query).change_context(Error::InvalidQuery)?;
 
-        Ok((validate_address(&contract)?, query))
+        let contract = contract
+            .parse::<AxelarAddress>()
+            .change_context(Error::InvalidContractAddress(contract))?;
+
+        Ok((contract, query))
     }
-}
-
-fn validate_address(address: &str) -> Result<TMAddress, Error> {
-    let address = TMAddress::from_str(address)
-        .change_context(Error::InvalidContractAddress(address.to_string()))?;
-    ensure!(
-        address.prefix() == PREFIX,
-        Error::InvalidContractAddress(address.to_string())
-    );
-
-    Ok(address)
 }
 
 impl Validate for Request<ContractsRequest> {
@@ -125,8 +117,8 @@ fn validate_key_id(key_id: KeyId) -> Result<(KeyIdString, tofnd::Algorithm), Err
 
 #[derive(Error, Debug)]
 pub enum Error {
-    #[error("empty filter")]
-    EmptyFilter,
+    #[error("invalid filter")]
+    InvalidFilter,
     #[error("invalid contract address {0}")]
     InvalidContractAddress(String),
     #[error("invalid query")]
@@ -143,97 +135,6 @@ pub enum Error {
     InvalidChainName(String),
 }
 
-#[derive(Debug)]
-pub enum EventFilter {
-    EventType(nonempty::String),
-    Contract(TMAddress),
-    EventTypeAndContract(nonempty::String, TMAddress),
-}
-
-impl TryFrom<ampd_proto::EventFilter> for EventFilter {
-    type Error = Report<Error>;
-
-    fn try_from(event_filter: ampd_proto::EventFilter) -> Result<Self, Error> {
-        let event_type = event_filter.r#type.try_into().ok();
-        let contract = if event_filter.contract.is_empty() {
-            None
-        } else {
-            Some(validate_address(&event_filter.contract)?)
-        };
-
-        match (event_type, contract) {
-            (Some(event_type), Some(contract)) => {
-                Ok(EventFilter::EventTypeAndContract(event_type, contract))
-            }
-            (Some(event_type), None) => Ok(EventFilter::EventType(event_type)),
-            (None, Some(contract)) => Ok(EventFilter::Contract(contract)),
-            (None, None) => Err(report!(Error::EmptyFilter)),
-        }
-    }
-}
-
-impl EventFilter {
-    pub fn filter(&self, event_type: &str, contract: Option<&TMAddress>) -> bool {
-        match self {
-            EventFilter::EventType(event_type_filter) => event_type_filter == event_type,
-            EventFilter::Contract(contract_filter) => Some(contract_filter) == contract,
-            EventFilter::EventTypeAndContract(event_type_filter, contract_filter) => {
-                event_type_filter == event_type && Some(contract_filter) == contract
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct EventFilters {
-    filters: Vec<EventFilter>,
-    include_block_begin_end: bool,
-}
-
-impl EventFilters {
-    pub fn filter(&self, event: &events::Event) -> bool {
-        let contract = event.contract_address();
-
-        match event {
-            events::Event::BlockBegin(_) | events::Event::BlockEnd(_) => {
-                self.include_block_begin_end
-            }
-            events::Event::Abci { event_type, .. } => self.filter_abci_event(event_type, contract),
-        }
-    }
-
-    fn filter_abci_event<T>(&self, event_type: &str, contract: Option<T>) -> bool
-    where
-        T: Into<TMAddress>,
-    {
-        if self.filters.is_empty() {
-            return true;
-        }
-
-        let contract = contract.map(Into::into);
-
-        self.filters
-            .iter()
-            .any(|filter| filter.filter(event_type, contract.as_ref()))
-    }
-}
-
-impl TryFrom<(Vec<ampd_proto::EventFilter>, bool)> for EventFilters {
-    type Error = Report<Error>;
-
-    fn try_from(
-        (event_filters, include_block_begin_end): (Vec<ampd_proto::EventFilter>, bool),
-    ) -> Result<Self, Error> {
-        Ok(EventFilters {
-            filters: event_filters
-                .into_iter()
-                .map(EventFilter::try_from)
-                .collect::<Result<_, _>>()?,
-            include_block_begin_end,
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::iter;
@@ -244,112 +145,7 @@ mod tests {
 
     use super::*;
     use crate::types::TMAddress;
-
-    #[test]
-    fn event_filter_should_be_created_from_valid_event_type() {
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "test_event".to_string(),
-            contract: "".to_string(),
-        };
-
-        let filter = EventFilter::try_from(proto_filter).unwrap();
-        assert!(matches!(filter, EventFilter::EventType(_)));
-    }
-
-    #[test]
-    fn event_filter_should_be_created_from_valid_contract_address() {
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "".to_string(),
-            contract: TMAddress::random(PREFIX).to_string(),
-        };
-
-        let filter = EventFilter::try_from(proto_filter).unwrap();
-        assert!(matches!(filter, EventFilter::Contract(_)));
-    }
-
-    #[test]
-    fn event_filter_should_be_created_from_valid_event_type_and_contract_address() {
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "test_event".to_string(),
-            contract: TMAddress::random(PREFIX).to_string(),
-        };
-
-        let filter = EventFilter::try_from(proto_filter).unwrap();
-        assert!(matches!(filter, EventFilter::EventTypeAndContract(_, _)));
-    }
-
-    #[test]
-    fn event_filter_should_fail_for_empty_filter() {
-        let proto_filter = ampd_proto::EventFilter::default();
-
-        let result = EventFilter::try_from(proto_filter);
-        assert_err_contains!(result, Error, Error::EmptyFilter);
-    }
-
-    #[test]
-    fn event_filter_should_fail_for_invalid_contract_address() {
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "".to_string(),
-            contract: "invalid_address".to_string(),
-        };
-
-        let result = EventFilter::try_from(proto_filter);
-        assert_err_contains!(result, Error, Error::InvalidContractAddress(_));
-    }
-
-    #[test]
-    fn event_filter_should_fail_for_contract_with_wrong_prefix() {
-        let address = TMAddress::random("wrong");
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "".to_string(),
-            contract: address.to_string(),
-        };
-
-        let result = EventFilter::try_from(proto_filter);
-        assert_err_contains!(result, Error, Error::InvalidContractAddress(_));
-    }
-
-    #[test]
-    fn event_filter_should_match_by_event_type() {
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "test_event".to_string(),
-            contract: "".to_string(),
-        };
-
-        let filter = EventFilter::try_from(proto_filter).unwrap();
-        assert!(filter.filter("test_event", None));
-        assert!(!filter.filter("other_event", None));
-    }
-
-    #[test]
-    fn event_filter_should_match_by_contract() {
-        let address = TMAddress::random(PREFIX);
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "".to_string(),
-            contract: address.to_string(),
-        };
-
-        let filter = EventFilter::try_from(proto_filter).unwrap();
-
-        assert!(filter.filter("any_event", Some(&address)));
-        assert!(!filter.filter("any_event", Some(&TMAddress::random(PREFIX))));
-        assert!(!filter.filter("any_event", None));
-    }
-
-    #[test]
-    fn event_filter_should_match_by_both_event_type_and_contract() {
-        let address = TMAddress::random(PREFIX);
-        let proto_filter = ampd_proto::EventFilter {
-            r#type: "test_event".to_string(),
-            contract: address.to_string(),
-        };
-
-        let filter = EventFilter::try_from(proto_filter).unwrap();
-
-        assert!(filter.filter("test_event", Some(&address)));
-        assert!(!filter.filter("other_event", Some(&address)));
-        assert!(!filter.filter("test_event", Some(&TMAddress::random(PREFIX))));
-    }
+    use crate::PREFIX;
 
     #[test]
     fn event_filters_should_be_created_from_valid_proto_filters() {
@@ -380,7 +176,7 @@ mod tests {
         });
 
         let result = req.validate();
-        assert_err_contains!(result, Error, Error::EmptyFilter);
+        assert_err_contains!(result, Error, Error::InvalidFilter);
     }
 
     #[test]
@@ -502,7 +298,7 @@ mod tests {
 
     #[test]
     fn validate_contract_state_should_extract_contract_and_query() {
-        let address = TMAddress::random(PREFIX);
+        let address = AxelarAddress::random();
         let query_json = serde_json::json!({"get_config": {}});
         let query_bytes = serde_json::to_vec(&query_json).unwrap();
