+++ ampd/src/grpc/testdata/reqs_errors_to_status.golden
@@ -1,7 +1,7 @@
 [
     (
         InvalidArgument,
-        EmptyFilter,
+        InvalidFilter,
     ),
     (
         InvalidArgument,
