+++ ampd/src/handlers/xrpl_multisig.rs
@@ -7,8 +7,8 @@ use cosmrs::tx::Msg;
 use cosmrs::Any;
 use cosmwasm_std::{HexBinary, Uint64};
 use error_stack::ResultExt;
-use events::try_from;
 use events::Error::EventTypeMismatch;
+use events::{try_from, EventType};
 use hex::encode;
 use multisig::msg::ExecuteMsg;
 use multisig::types::MsgToSign;
@@ -20,6 +20,7 @@ use tracing::info;
 use xrpl_types::types::XRPLAccountId;
 
 use crate::event_processor::EventHandler;
+use crate::event_sub::event_filter::{EventFilter, EventFilters};
 use crate::handlers::errors::Error::{self, DeserializeEvent};
 use crate::tofnd::{Algorithm, Multisig};
 use crate::types::{PublicKey, TMAddress};
@@ -179,6 +180,16 @@ where
                 Ok(vec![])
             }
         }
+    }
+
+    fn event_filters(&self) -> EventFilters {
+        EventFilters::new(
+            vec![EventFilter::EventTypeAndContract(
+                SigningStartedEvent::event_type(),
+                self.multisig.clone(),
+            )],
+            true,
+        )
     }
 }
